+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
+
+## next
+* Added patching of ES6 Promise to avoid context switch when accessing shared data source
+
 ## v1.0.2
 * Added child request local to parent if any exists to allow propagation of errors to the parent domain